pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = ""
authors = ["Wasim Rashid <wasim.rashid@developers.studio>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlmodel = "^0.0.16"
fastapi = "^0.110.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
